SECURITY: Adjustments regarding session cookie handling
Combined effect: unless a third-party can inject a cookie with a
random session id that is logged-in session, we disregard whatever
is given in the session and create a new one.

While technically if a person visits / with a non-logged-in valid
session (which is only possible if the user visits /oauthinit then
go back to /, or somehow the oauth tokens are invalidated), for that
very request to / the session id is the old one. However, the session
is immediately invalidated and the user is now forced to login with
/oauthinit, which refreshes the session ID.

Weird behavior may occur if the `v2c-session` cookie is actively
shadowed, and we cannot set the session cookie appropriately, in
which case user may be simply unable to login.

https://www.usenix.org/system/files/conference/usenixsecurity15/sec15-paper-zheng.pdf