Make the URIs configurable in config.json
The public ones, for defining the uris, are now also exposed to the
web and available to v2c.js.

v2c.js will no longer attempt to connect to socketio if the uri is
undefined, allowing a test instance that do not load the status from
the main instance.